rename newly introduced modules to make them private symbolically
